requirements.txt
requests
Pillow
numpy
selenium
browser-cookie3
scikit-learn
